refactor Role DTOs; reorganize package structure for Request and Response DTOs, update RoleController to include authentication headers

+++ src/main/java/com/supplychainx/service_user/controller/RoleController.java
@@ -1,8 +1,8 @@
 package com.supplychainx.service_user.controller;
 
 import com.supplychainx.handler.GlobalSuccessHandler;
-import com.supplychainx.service_user.dto.RoleRequestDTO;
-import com.supplychainx.service_user.dto.RoleResponseDTO;
+import com.supplychainx.service_user.dto.Request.RoleRequestDTO;
+import com.supplychainx.service_user.dto.Response.RoleResponseDTO;
 import com.supplychainx.service_user.service.RoleService;
 import com.supplychainx.util.AuthUtil;
 import jakarta.validation.Valid;
@@ -21,7 +21,10 @@ public class RoleController {
     private final AuthUtil authUtil;
 
     @GetMapping
-    public List<RoleResponseDTO> getAllRoles() {
+    public List<RoleResponseDTO> getAllRoles(
+            @RequestHeader("email") String email,
+            @RequestHeader("password") String password) {
+        authUtil.verifyAccess(email, password, "ADMIN");
         return roleService.getAll();
     }
 
@@ -36,18 +39,30 @@ public class RoleController {
     }
 
     @GetMapping("/{id}")
-    public RoleResponseDTO getRoleById(@PathVariable("id") Long id) {
+    public RoleResponseDTO getRoleById(
+            @RequestHeader("email") String email,
+            @RequestHeader("password") String password,
+            @PathVariable("id") Long id) {
+        authUtil.verifyAccess(email, password, "ADMIN");
         return roleService.getById(id);
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<Map<String,Object>> updateRole(@PathVariable("id") Long id, @Valid @RequestBody RoleRequestDTO roleRequestDTO) {
+    public ResponseEntity<Map<String, Object>> updateRole(
+            @RequestHeader("email") String email,
+            @RequestHeader("password") String password,
+            @PathVariable("id") Long id, @Valid @RequestBody RoleRequestDTO roleRequestDTO) {
+        authUtil.verifyAccess(email, password, "ADMIN");
         RoleResponseDTO updatedDTO = roleService.update(id, roleRequestDTO);
         return GlobalSuccessHandler.handleSuccessWithData("Role updated successfully", updatedDTO);
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Map<String, Object>> deleteRole(@PathVariable("id") Long id) {
+    public ResponseEntity<Map<String, Object>> deleteRole(
+            @RequestHeader("email") String email,
+            @RequestHeader("password") String password,
+            @PathVariable("id") Long id) {
+        authUtil.verifyAccess(email, password, "ADMIN");
         roleService.delete(id);
         return GlobalSuccessHandler.handleDeleted("Role deleted successfully");
     }
